style: :art: Transfert simulate_transactions utils from test_simulate_transactions.rs to lib.rs
Transfert simulate_transactions utils from test_simulate_transactions.rs to lib.rs for a better code format
